JBIDE-18247 : delay remote variable resolution to when necessary
When displaying Central, ProxyWizardManager and RefreshBuzzJob instantiations trigger fetching of remote resources (ide-config.property), that could block the UI for an indefinite amount of time.
The values being looked up are not required until later, by methods running in Jobs, so the current fix consists in performing the lookups only when necessary and when it's safe to do it.

The fix is certainly not ideal, other parts of the code calling PropertiesHelper.getPropertiesProvider().getValue(...) might require better handling as well, but this would require major refactoring at the foundation level (introducing/handling timeouts).

Until a better redesign can be performed, the current fix makes Central not blocking the UI, as expected. Some jobs keeps running for a very long time in the background but they can be cancelled any time.

Signed-off-by: Fred Bricon <fbricon@gmail.com>