JBIDE-28815 - adapt to internal API break from m2e
a properties method parameter has been switched to a Map. This
implementation propose to do the conversion when needed without touching
other parts of the codebase.

Another alternative could be to no more use the Properties class and
always use the map, It would avoid convsersions. It seems that the
Properties structure is not really leveraged elsewhere BUT it would
introduce an API break.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>