JBIDE-28815: fix after TP update (#499)
* Fix mockito dependency after upgrade, fixes JBIDE-28815

Signed-off-by: Ondrej Dockal <odockal@redhat.com>

* JBIDE-28815 - adapt to internal API break from m2e

a properties method parameter has been switched to a Map. This
implementation propose to do the conversion when needed without touching
other parts of the codebase.

Another alternative could be to no more use the Properties class and
always use the map, It would avoid convsersions. It seems that the
Properties structure is not really leveraged elsewhere BUT it would
introduce an API break.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>

* JBIDE-28815: fix API breakeage

Signed-off-by: Stephane Bouchet <sbouchet@redhat.com>

* JBIDE-28815: bump version

Signed-off-by: Stephane Bouchet <sbouchet@redhat.com>

Signed-off-by: Ondrej Dockal <odockal@redhat.com>
Signed-off-by: Aurélien Pupier <apupier@redhat.com>
Signed-off-by: Stephane Bouchet <sbouchet@redhat.com>
Co-authored-by: Ondrej Dockal <odockal@redhat.com>
Co-authored-by: Aurélien Pupier <apupier@redhat.com>